refactor: vo, entity, enums로 나누어진 패키지를 model로 통합
