+++ src/main/java/com/goggles/orderservice/application/dto/command/CreateOrderItemCommand.java
@@ -1,6 +1,6 @@
 package com.goggles.orderservice.application.dto.command;
 
-import com.goggles.orderservice.domain.vo.Instructor;
-import com.goggles.orderservice.domain.vo.Product;
+import com.goggles.orderservice.domain.model.Instructor;
+import com.goggles.orderservice.domain.model.Product;
 
 public record CreateOrderItemCommand(Product product, Instructor instructor) {}
