+++ src/main/java/com/goggles/orderservice/domain/model/OrderItemType.java
@@ -1,4 +1,4 @@
-package com.goggles.orderservice.domain.enums;
+package com.goggles.orderservice.domain.model;
 
 public enum OrderItemType {
   COURSE,
