+++ src/main/java/com/goggles/orderservice/domain/model/OrderSortType.java
@@ -1,4 +1,4 @@
-package com.goggles.orderservice.domain.enums;
+package com.goggles.orderservice.domain.model;
 
 import jakarta.ws.rs.BadRequestException;
 import java.util.Arrays;
