+++ src/test/java/com/goggles/orderservice/application/OrderQueryTest.java
@@ -15,17 +15,17 @@ import com.goggles.orderservice.application.dto.result.OrderDetailResult;
 import com.goggles.orderservice.application.dto.result.OrderItemSummary;
 import com.goggles.orderservice.application.dto.result.OrderListResult;
 import com.goggles.orderservice.application.service.impl.OrderQueryServiceImpl;
-import com.goggles.orderservice.domain.entity.Order;
-import com.goggles.orderservice.domain.enums.OrderItemStatus;
-import com.goggles.orderservice.domain.enums.OrderItemType;
-import com.goggles.orderservice.domain.enums.OrderStatus;
+import com.goggles.orderservice.domain.model.Order;
+import com.goggles.orderservice.domain.model.OrderItemStatus;
+import com.goggles.orderservice.domain.model.OrderItemType;
+import com.goggles.orderservice.domain.model.OrderStatus;
 import com.goggles.orderservice.domain.repository.OrderPageQuery;
 import com.goggles.orderservice.domain.repository.OrderRepository;
-import com.goggles.orderservice.domain.vo.Instructor;
-import com.goggles.orderservice.domain.vo.OrderItemSpec;
-import com.goggles.orderservice.domain.vo.OrderPrice;
-import com.goggles.orderservice.domain.vo.Orderer;
-import com.goggles.orderservice.domain.vo.Product;
+import com.goggles.orderservice.domain.model.Instructor;
+import com.goggles.orderservice.domain.model.OrderItemSpec;
+import com.goggles.orderservice.domain.model.OrderPrice;
+import com.goggles.orderservice.domain.model.Orderer;
+import com.goggles.orderservice.domain.model.Product;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
